### 2025-09-19 #9012 @editorialbot
[PRE REVIEW]: rcldf: R library for reading CLDF files
**Submitting author:** @SimonGreenhill (<a href="http://orcid.org/0000-0001-7832-6156">Simon Greenhill</a>)
**Repository:** https://github.com/SimonGreenhill/rcldf
**Branch with paper.md** (empty if default branch):
**Version:** v1.5.1
