+++ DESCRIPTION
@@ -5,12 +5,13 @@ Version: 1.5.1
 Authors@R: 
     person(given = c("Simon", "J."), family = "Greenhill", email = "simon@simon.net.nz", role = c("aut", "cre"))
 Maintainer: Simon J. Greenhill <simon@simon.net.nz>
-Description: The rcldf package is designed to facilitate the manipulation and analysis of datasets in
-    Cross-Linguistic Data Format (CLDF, Forkel et al. 2018 <doi:10.1038/sdata.2018.205>). CLDF is a 
-    framework for storing cross-linguistic data, ensuring compatibility and ease of data
-    exchange between different linguistic datasets. This package simplifies the loading, querying,
-    and visualisation of CLDF datasets making it easier to conduct comparative linguistic analyses,
-    manage language data, and apply statistical methods directly within R.
+Description: Cross-Linguistic Data Format (CLDF) is a framework for storing 
+    cross-linguistic data, ensuring compatibility and ease of data exchange between
+    different linguistic datasets see Forkel et al. (2018) <doi:10.1038/sdata.2018.205>. 
+    The 'rcldf' package is designed to facilitate the manipulation and analysis of 
+    these datasets by simplifying the loading, querying, and visualisation of CLDF
+    datasets making it easier to conduct comparative linguistic analyses, manage 
+    language data, and apply statistical methods directly within R.
 License: Apache License (>= 2.0)
 Encoding: UTF-8
 Imports:
