+++ tests/testthat/test_cldf.R
@@ -68,9 +68,32 @@ test_that("test handling of no sources", {
 })
 
 
+test_that("test handling of no sources", {
+    df <- cldf(system.file("extdata/examples/no_sources", package = "rcldf"))
+    expect_equal(is.na(df$sources), TRUE)
+
+    out <- capture.output(summary(df))
+    expect_match(out[6], "Sources: 0")
+})
+
+
 test_that("test handling of valid/invalid JSON files", {
     expect_error(
         cldf(system.file("extdata/examples/not_a_cldf/also_not_a_cldf/invalid.json", package = "rcldf")),
         "Metadata doesn't define any tables, or a url to build a table definition from"
     )
+})
+
+
+
+
+
+#skip=dialect$headerRowCount,  # causes more problems than it's worth?
+#col_names=column_names,
+test_that("Error with dialect$headerRowCount and skip", {
+    o <- cldf('datasets/barlownumerals')
+    p <- vroom::problems(o$tables$ParameterTable, "problems")
+    expect_equal(nrow(p), 0)
+    # first row should NOT be the header row repeated ("ID", "Name", ...)
+    expect_equal(as.character(o$tables$ParameterTable[1, 'ID']), 'numeral-system')
 })
