### 2026-01-20 #9012 @SimonGreenhill
Thanks @arfon. Quick question. @samhforbes queried whether this was substantial enough for JOSS. Is it worth me adding these sections or will the package be out of scope?
