### 2026-02-12 d7df13a
add load_dplace
