+++ man/load_dplace.Rd
@@ -1,0 +1,19 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/load_glottolog.R
+\name{load_dplace}
+\alias{load_dplace}
+\title{Returns a CLDF dataset object of the latest D-PLACE version.}
+\usage{
+load_dplace(load_bib = FALSE, cache_dir = NULL)
+}
+\arguments{
+\item{load_bib}{load sources (TRUE/FALSE, default FALSE)}
+
+\item{cache_dir}{A cache_dir to use. If NULL it will use get_cache_dir}
+}
+\value{
+A \code{cldf} object
+}
+\description{
+Returns a CLDF dataset object of the latest D-PLACE version.
+}
