+++ DESCRIPTION
@@ -29,7 +29,8 @@ Imports:
     rlang,
     tools,
     urltools,
-    utils
+    utils,
+    vroom
 Suggests: 
     ggplot2,
     patchwork,

[…]
