### 2026-02-15 #9012 @arfon
> Thanks @arfon. Quick question. @samhforbes queried whether this was substantial enough for JOSS. Is it worth me adding these sections or will the package be out of scope?

The software passed an initial triage but to make a complete assessment we need the new paper sections (as this will now be assessed against the updated 2026 scope). Sorry I can't be more certain.
