### 2026-03-04 d31bb8a
add get_foreign_keys
