+++ NAMESPACE
@@ -5,6 +5,7 @@ S3method(print,cldf_schema)
 S3method(summary,cldf)
 export(as.cldf.wide)
 export(cldf)
+export(datasets)
 export(datatype_to_type)
 export(get_cache_dir)
 export(get_details)
@@ -27,7 +28,9 @@ export(resolve_path)
 export(schema)
 export(separate)
 export(set_cache_dir)
+export(subset_cldf)
 importFrom(digest,digest)
+importFrom(dplyr,relocate)
 importFrom(jsonlite,fromJSON)
 importFrom(magrittr,"%>%")
 importFrom(purrr,lmap)
