+++ tests/testthat/test_is_url.R
@@ -7,4 +7,6 @@ test_that("test is_url", {
     expect_equal(is_url("https://zenodo.org/records/10901700/files/huon_test_data-0.01.zip?download=1"), TRUE)
     expect_equal(is_url("gopher://simon.net.nz/something/file.zip"), TRUE)  # why not?
     expect_equal(is_url("/home/simon"), FALSE)
+
+    expect_equal(is_url("HTTP://SIMON.NET.NZ"), TRUE)
 })
