+++ man/subset_cldf.Rd
@@ -1,0 +1,16 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/subset_cldf.R
+\name{subset_cldf}
+\alias{subset_cldf}
+\title{Subset a CLDF object with Cascading Filters}
+\usage{
+subset_cldf(x, expr)
+}
+\arguments{
+\item{x}{A cldf object.}
+
+\item{expr}{A logical expression (e.g., Language_ID == 'kate')}
+}
+\description{
+Subset a CLDF object with Cascading Filters
+}
