+++ man/load_dataset.Rd
@@ -1,0 +1,43 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/datasets.R
+\name{load_dataset}
+\alias{load_dataset}
+\title{Load a CLDF dataset by name and version}
+\usage{
+load_dataset(dataset, version = NULL, source = "Zenodo")
+}
+\arguments{
+\item{dataset}{a character string naming the dataset (must match the
+\code{Dataset} column in \code{\link{datasets}}).}
+
+\item{version}{a character string specifying the version to load (e.g.
+\code{"v1.4.1"}). Defaults to \code{NULL}, which selects the latest
+available version.}
+
+\item{source}{a character string, either \code{"Zenodo"} (default) or
+\code{"GitHub"}, specifying where to download the dataset from.
+Zenodo downloads are recommended as they are archival and stable.}
+}
+\value{
+A \code{cldf} object.
+}
+\description{
+Looks up a dataset from the registry returned by \code{\link{datasets}},
+resolves the requested version, and downloads it from either Zenodo or
+GitHub.
+}
+\examples{
+\dontrun{
+# load the latest version of a dataset
+ds <- load_dataset("vanuatuvoices")
+
+# load a specific version
+ds <- load_dataset("vanuatuvoices", version = "v1.3")
+
+# load from GitHub instead
+ds <- load_dataset("vanuatuvoices", source = "GitHub")
+}
+}
+\seealso{
+\code{\link{datasets}}, \code{\link{cldf}}
+}
