+++ R/leaflet.R
@@ -1,0 +1,121 @@
+#' Interactive Leaflet Map for CLDF
+#'
+#' @param x A cldf object.
+#' @param color_by Column to color by (e.g., "Family").
+#' @export
+plot_languages <- function(x, color_by = "ID") {
+    if (!requireNamespace("leaflet", quietly = TRUE)) {
+        stop("Package 'leaflet' is required for interactive maps. Please install it.")
+    }
+
+    lt <- x$tables$LanguageTable
+    lt <- lt[!is.na(lt$Latitude) & !is.na(lt$Longitude), ]
+    lt$Longitude[lt$Longitude < 0] <- lt$Longitude[lt$Longitude < 0] + 360
+
+
+    # Define a vibrant color palette
+    pal_fn <- leaflet::colorFactor(palette = "Set1", domain = lt[[color_by]])
+
+    leaflet::leaflet(lt, options = leaflet::leafletOptions(worldCopyJump = TRUE)) %>%
+        # Use the "Positron" theme for a clean, pure-white background
+        leaflet::addProviderTiles("CartoDB.Positron") %>%
+        leaflet::addCircleMarkers(
+            lng = ~Longitude, lat = ~Latitude,
+            color = "white", weight = 1,
+            fillColor = ~pal_fn(lt[[color_by]]),
+            fillOpacity = 0.8,
+            radius = 6,
+            popup = ~paste0("<b>", Name, "</b><br>", color_by, ": ", lt[[color_by]]),
+            label = ~Name
+        ) %>%
+        leaflet::addLegend(pal = pal_fn, values = ~lt[[color_by]], title = color_by)
+}
+
+
+
+
+plot_parameter <- function(x, parameter = "1sg_a", color_by = 'Value') {
+    if (!requireNamespace("leaflet", quietly = TRUE)) {
+        stop("Package 'leaflet' is required for interactive maps. Please install it.")
+    }
+
+    # does param exist?
+    if (parameter %in% x$tables$ParameterTable$ID == FALSE) {
+        stop("Invalid Parameter_ID")
+    }
+    # do we have form table or value table
+    fk <- rcldf:::get_foreign_keys(x)
+    fk <- fk[fk$SourceColumn == 'Parameter_ID', ]
+
+    df <- as.cldf.wide(x, x$resources[[ fk[['SourceTable']][[1]] ]])
+    df <- df[df$Parameter_ID == parameter,]
+
+    # remove no locations and standardise
+    df <- df[! is.na(df$Longitude) & ! is.na(df$Latitude), ]
+    df$Longitude[df$Longitude < 0] <- df$Longitude[df$Longitude < 0] + 360
+
+    pal_fn <- leaflet::colorFactor(palette = "Set1", domain = df[[color_by]])
+
+    leaflet::leaflet(df, options = leaflet::leafletOptions(worldCopyJump = TRUE)) |>
+        leaflet::addProviderTiles("CartoDB.Positron") |>
+        leaflet::addCircleMarkers(
+            lng = ~Longitude, lat = ~Latitude,
+            color = "white", weight = 1,
+            fillColor = ~pal_fn(df[[color_by]]),
+            fillOpacity = 0.8,
+            radius = 6,
+            popup = ~paste0("<b>", Name.LanguageTable, "</b><br>", color_by, ": ", df[[color_by]]),
+            label = ~Name.LanguageTable
+        ) %>%
+        leaflet::addLegend(pal = pal_fn, values = ~df[[color_by]], title = color_by)
+
+}
+
+
+
+
+
+plot_word <- function(x, parameter = "1sg_a", color_by = 'Cognacy') {
+    if (!requireNamespace("leaflet", quietly = TRUE)) {
[…]
