+++ NAMESPACE
@@ -19,10 +19,12 @@ export(is_url)
 export(list_cache_files)
 export(load_clts)
 export(load_concepticon)
+export(load_dataset)
 export(load_dplace)
 export(load_glottolog)
 export(make_cache_key)
 export(nullify)
+export(plot_languages)
 export(read_bib)
 export(resolve_path)
 export(schema)
@@ -30,7 +32,6 @@ export(separate)
 export(set_cache_dir)
 export(subset_cldf)
 importFrom(digest,digest)
-importFrom(dplyr,relocate)
 importFrom(jsonlite,fromJSON)
 importFrom(magrittr,"%>%")
 importFrom(purrr,lmap)
@@ -41,3 +42,4 @@ importFrom(stats,setNames)
 importFrom(urltools,domain)
 importFrom(urltools,path)
 importFrom(utils,unzip)
+importFrom(versionsort,ver_latest)
