+++ man/plot_languages.Rd
@@ -1,0 +1,16 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/leaflet.R
+\name{plot_languages}
+\alias{plot_languages}
+\title{Interactive Leaflet Map for CLDF}
+\usage{
+plot_languages(x, color_by = "ID")
+}
+\arguments{
+\item{x}{A cldf object.}
+
+\item{color_by}{Column to color by (e.g., "Family").}
+}
+\description{
+Interactive Leaflet Map for CLDF
+}
