### 2026-03-18 fa0acca
add tests
