+++ tests/testthat/test_resolve_path.R
@@ -61,3 +61,52 @@ test_that("resolve_path handles archives (.zip)", {
     # run again, should message
     expect_message(resolve_path(zfile, tempdir()), "^Reusing cache in")
 })
+
+
+test_that("resolve_path creates cache dir", {
+    expected <- csvwr::read_metadata(MD_JSON_PATH)
+    tmp <- file.path(tempdir(), 'resolve_path_creates_cache_dir')
+    obtained <- resolve_path(MD_JSON_PATH, tmp)
+    expect_equal(obtained$metadata, expected)
+})
+
+
+test_that("resolve_path handles cache hit", {
+    path = "https://example.com/test"
+    tmp <- tempdir()
+    cache_key_dir <- file.path(tmp, make_cache_key(path))
+    # Create the folder that make_cache_key would point to
+    dir.create(cache_key_dir, recursive = TRUE)
+    # We expect it to fail later because the folder is empty (no .json),
+    # but the logic must pass through the 'cache hit' branch.
+    expect_error(resolve_path(path, cache_dir = tmp), "no metadata JSON file found")
+})
+
+
+test_that("resolve_path handles github", {
+    with_mocked_bindings({
+        expect_error(
+            resolve_path("https://github.com/org/repo"),
+            "Path .* does not exist"  # crash and burn
+        )
+    },
+    remote_download = function(...) "mock_remote_download",
+    .package = "remotes"
+    )
+})
+
+
+test_that("resolve_path handles generic zip files", {
+    with_mocked_bindings({
+        expect_error(
+            resolve_path("https://example.com/filename.zip?junk"),
+            "no metadata JSON file found in"
+        )
+    },
+    download.file = function(url, dest, ...) {
+        file.create(dest)
+        return(0)
+    },
+    .package = 'utils'
+    )
+})
