+++ paper/paper.bib
@@ -207,21 +207,135 @@
 }
 
 @misc{AUTOTYP,
-  doi = {10.5281/ZENODO.7976754},
-  url = {https://zenodo.org/record/7976754},
-  author = {Bickel, Balthasar and Nichols, Johanna and Zakharko, Taras and Witzlack-Makarevich, Alena and Hildebrandt, Kristine and Rießler, Michael and Bierkandt, Lennart and Zúñiga, Fernando and Lowe, John B.},
-  keywords = {database, linguistics, linguistic typology, morphosyntax, syntax, morphology, phonology},
-  title = {The AUTOTYP database},
-  publisher = {Zenodo},
-  year = {2023},
-  copyright = {Creative Commons Attribution 4.0 International}
+  title        = {The AUTOTYP database},
+  author       = {Bickel, Balthasar and Nichols, Johanna and Zakharko, Taras and Witzlack-Makarevich, Alena and Hildebrandt, Kristine and Rießler, Michael and Bierkandt, Lennart and Zúñiga, Fernando and Lowe, John B.},
+  year         = 2023,
+  publisher    = {Zenodo},
+  doi          = {10.5281/ZENODO.7976754},
+  url          = {https://zenodo.org/record/7976754},
+  copyright    = {Creative Commons Attribution 4.0 International},
+  keywords     = {database, linguistics, linguistic typology, morphosyntax, syntax, morphology, phonology}
 }
 
 @book{EWAVE,
-  address   = {},
-  editor    = {Bernd Kortmann and Kerstin Lunkenheimer and Katharina Ehret},
-  publisher = {},
-  title     = {eWAVE},
-  url       = {https://ewave-atlas.org/},
-  year      = {2020}
+  title        = {eWAVE},
+  year         = 2020,
+  publisher    = {},
+  address      = {},
+  url          = {https://ewave-atlas.org/},
+  editor       = {Bernd Kortmann and Kerstin Lunkenheimer and Katharina Ehret}
+}
+
+@article{Ziemann_2016,
+  title        = {Gene name errors are widespread in the scientific literature},
+  author       = {Ziemann, Mark and Eren, Yotam and El-Osta, Assam},
+  year         = 2016,
+  month        = aug,
+  journal      = {Genome Biology},
+  publisher    = {Springer Science and Business Media LLC},
+  volume       = 17,
+  number       = 1,
+  doi          = {10.1186/s13059-016-1044-7},
+  issn         = {1474-760X},
+  url          = {http://dx.doi.org/10.1186/s13059-016-1044-7}
+}
+
+@inproceedings{Forkel2024,
+  title        = {Converting Legacy Data to {CLDF}: A {FAIR} Exit Strategy for Linguistic Web Apps},
+  author       = {Forkel, Robert  and Swanson, Daniel  and Moran, Steven},
+  year         = 2024,
+  month        = may,
+  booktitle    = {Proceedings of the 2024 Joint International Conference on Computational Linguistics, Language Resources and Evaluation (LREC-COLING 2024)},
+  publisher    = {ELRA and ICCL},
+  address      = {Torino, Italia},
+  pages        = {3978--3982},
+  url          = {https://aclanthology.org/2024.lrec-main.353/},
+  editor       = {Calzolari, Nicoletta  and Kan, Min-Yen  and Hoste, Veronique  and Lenci, Alessandro  and Sakti, Sakriani  and Xue, Nianwen},
[…]
