+++ paper/paper.bib
@@ -339,3 +339,52 @@
   number = {4},
   pages = {590--621}
 }
+
+
+@article{Kortmann2021,
+  title = {Reflecting on the Quantitative Turn in Linguistics},
+  author = {Kortmann, Bernd},
+  date = {2021},
+  journaltitle = {Linguistics},
+  volume = {59},
+  number = {5},
+  pages = {1207--1226},
+  doi = {doi:10.1515/ling-2019-0046},
+  url = {https://doi.org/10.1515/ling-2019-0046},
+  urldate = {2026-03-18}
+}
+
+
+@article{McGillivray2022,
+  title = {Deep {{Impact}}: {{A Study}} on the {{Impact}} of {{Data Papers}} and {{Datasets}} in the {{Humanities}} and {{Social Sciences}}},
+  shorttitle = {Deep {{Impact}}},
+  author = {McGillivray, Barbara and Marongiu, Paola and Pedrazzini, Nilo and Ribary, Marton and Wigdorowitz, Mandy and Zordan, Eleonora},
+  date = {2022-12},
+  journaltitle = {Publications},
+  volume = {10},
+  number = {4},
+  pages = {39},
+  publisher = {Multidisciplinary Digital Publishing Institute},
+  issn = {2304-6775},
+  doi = {10.3390/publications10040039},
+  url = {https://www.mdpi.com/2304-6775/10/4/39},
+  urldate = {2026-03-18},
+  abstract = {The humanities and social sciences (HSS) have recently witnessed an exponential growth in data-driven research. In response, attention has been afforded to datasets and accompanying data papers as outputs of the research and dissemination ecosystem. In 2015, two data journals dedicated to HSS disciplines appeared in this landscape: Journal of Open Humanities Data (JOHD) and Research Data Journal for the Humanities and Social Sciences (RDJ). In this paper, we analyse the state of the art in the landscape of data journals in HSS using JOHD and RDJ as exemplars by measuring performance and the deep impact of data-driven projects, including metrics (citation count; Altmetrics, views, downloads, tweets) of data papers in relation to associated research papers and the reuse of associated datasets. Our findings indicate: that data papers are published following the deposit of datasets in a repository and usually following research articles; that data papers have a positive impact on both the metrics of research papers associated with them and on data reuse; and that Twitter hashtags targeted at specific research campaigns can lead to increases in data papers’ views and downloads. HSS data papers improve the visibility of datasets they describe, support accompanying research articles, and add to transparency and the open research agenda.},
+  langid = {english},
+  keywords = {data journals,data papers,data reuse,humanities,impact,open data,open humanities,open research,social sciences}
+}
+
+
+@incollection{Tresoldi2022,
+  author = {Tresoldi, Tiago  and Rzymski, Christoph and Forkel, Robert and Greenhill, Simon J. and List, Johann-Mattis and Gray,  Russell D.},
+  editor = {Berez-Kroeker, Andrea L. and McDonnell, Bradley  and Koller, Eve and Collister, 
+Lauren B.},
+  doi = {10.7551/mitpress/12200.003.0033},
+  booktitle = {The Open Handbook of Linguistic Data Management},
+  pages = {345--354},
+  title = {{Managing Historical Linguistic Data for Computational Phylogenetics and Computer-Assisted Language Comparison}},
+  year = {2022},
+  abstract = {Computational phylogenetics is a relatively recent branch of historical linguistics that uses quantitative techniques to investigate the history of related languages. As the classical comparative method is less explicit on the techniques for constructing phylogenies of language families (see discussion in Jacques & List 2019), such a new approach can complement traditional techniques for sub-grouping based on shared innovations (Ross & Durie 1996).},
+  publisher = {MIT Press},
+  address = {Massachusetts}
+}
