+++ paper/paper.md
@@ -20,28 +20,44 @@ bibliography: paper.bib
 
 # Summary
 
-Cross-Linguistic Data Formats (CLDF) is a standardized data format becoming increasingly common for storing and 
-distributing a wide range of comparative linguistic, cultural, ethnographic, geographic, and religious data.
-The `rcldf` package provides a lightweight _R_ toolkit for loading and reading CLDF files from both local and remote sources.
-The package facilitates analysis with _R_ by providing a number of convenience methods for converting CLDF data, and connecting to standard "reference catalogues". The aim of `rcldf` is to provide researchers with a robust toolkit for 
-seamlessly integrating CLDF datasets into their workflows, enhancing the efficiency of linguistic and cultural research.
-
+Cross-Linguistic Data Formats (CLDF) is a standardized data format becoming increasingly common for storing
+and distributing a wide range of comparative linguistic, cultural, ethnographic, geographic, and religious
+data. The `rcldf` package provides a lightweight _R_ toolkit for loading and reading CLDF files from both
+local and remote sources. The package facilitates analysis with _R_ by providing a number of convenience
+methods for converting CLDF data, and connecting to standard 'reference catalogues'. The aim of `rcldf` is
+to provide researchers with a robust toolkit for seamlessly integrating CLDF datasets into their workflows,
+enhancing the efficiency of linguistic and cultural research.
 
 # Statement of need
 
 Cross-Linguistic Data Formats [CLDF, @Forkel2018] is a standardized data
-format designed to handle cross-linguistic and cross-cultural datasets.
-CLDF provides a consistent specification and package format ([https://cldf.clld.org/](https://cldf.clld.org/)) for common types of 
-linguistic and cultural data from word lists, to grammatical features, and cultural
-traits. The aim of CLDF is to provide a simple, reliable data format to facilitate
-the storage, sharing, and re-use for these data. 
+format designed to handle cross-linguistic and cross-cultural datasets. 
+There are currently almost [300 CLDF datasets
+available](https://zenodo.org/search?q=CLDF&f=resource_type%3Adataset&l=list&p=1&s=10&sort=bestmatch),
+representing many of the major publically available linguistic and cultural datasets. 
+These datasets contain a wide variety of different types of data from the
+world's languages and cultures including catalogues of linguistic metadata, 
+to lexical word lists, grammatical features, phonetic information, 
+geographical locations and areas, as well as religious and cultural traits (Table 1).
 
-There are currently more than [250 CLDF datasets
-available](https://zenodo.org/search?q=CLDF&f=resource_type%3Adataset&l=list&p=1&s=10&sort=bestmatch)
-containing data from the world's languages and cultures including
-everything from catalogues of linguistic metadata, to word lists of lexical data, 
-grammatical features, phonetic information, geographic information, and
-religious and cultural databases (Table 1).
+CLDF provides a consistent specification and package format ([https://cldf.clld.org/](https://cldf.clld.org/)). This format is a lightweight data-package
+containing one or more data tables containing tabular data in 'CSV on the Web' format (CSVW)
+following the World Wide Web Consortium (W3C) recommendations for [Tabular Data and Metadata](https://www.w3.org/TR/tabular-data-model/). These tables are described and connected by a metadata file in [Javascript Object Notation](http://json.org/) (JSON) format that details the datatypes
+in the CSVW files and how they are related to each other, and to global 'reference catalogs' of 
+linguistic and cultural information [@Forkel2018]. 
+
+CLDF is rapidly becoming a standard framework for storing cultural data as it provides a simple, 
+reliable data format that facilitate the storage, sharing, and re-use for these data. The format
[…]
