+++ man/plot_parameter.Rd
@@ -1,0 +1,23 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/leaflet.R
+\name{plot_parameter}
+\alias{plot_parameter}
+\title{Plot Distribution of a Specific Parameter}
+\usage{
+plot_parameter(x, parameter = "1sg_a", color_by = "Value")
+}
+\arguments{
+\item{x}{A cldf object.}
+
+\item{parameter}{Character string. The ID of the parameter to plot (e.g., "1sg_a").}
+
+\item{color_by}{Character string. The column to use for the color scale (e.g., "Value").}
+}
+\value{
+A leaflet map object.
+}
+\description{
+Filters the dataset for a specific Parameter ID and maps the values across languages.
+This function automatically resolves whether the data is in a Form or Value table
+and joins it with geographic data.
+}
