+++ man/plot_word.Rd
@@ -1,0 +1,23 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/leaflet.R
+\name{plot_word}
+\alias{plot_word}
+\title{Plot Words/Forms as Text Labels on a Map}
+\usage{
+plot_word(x, parameter = "1sg_a", color_by = "Cognacy")
+}
+\arguments{
+\item{x}{A cldf object.}
+
+\item{parameter}{Character string. The ID of the parameter (word) to plot.}
+
+\item{color_by}{Character string. Column used to categorize and color the text labels.}
+}
+\value{
+A leaflet map object.
+}
+\description{
+Similar to \code{plot_parameter}, but instead of circles, this function renders the
+actual phonetic forms (Value) as text labels directly on the map. Labels are
+color-coded based on the \code{color_by} column (e.g., Cognacy).
+}
