+++ NAMESPACE
@@ -25,6 +25,8 @@ export(load_glottolog)
 export(make_cache_key)
 export(nullify)
 export(plot_languages)
+export(plot_parameter)
+export(plot_word)
 export(read_bib)
 export(resolve_path)
 export(schema)
