+++ R/get_foreign_keys.R
@@ -1,4 +1,11 @@
-
+#' Returns a table of the foreign keys in a CLDF dataset.
+#'
+#' @param cldf_obj a CLDF object
+#' @return a dataframe
+#' @export
+#' @examples
+#' o <- cldf(system.file("extdata/huon", "cldf-metadata.json", package = "rcldf"))
+#' get_foreign_keys(o)
 get_foreign_keys <- function(cldf_obj) {
     results <- lapply(cldf_obj$metadata$tables, function(table) {
         fks <- table$tableSchema$foreignKeys
