+++ man/get_foreign_keys.Rd
@@ -1,0 +1,21 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/get_foreign_keys.R
+\name{get_foreign_keys}
+\alias{get_foreign_keys}
+\title{Returns a table of the foreign keys in a CLDF dataset.}
+\usage{
+get_foreign_keys(cldf_obj)
+}
+\arguments{
+\item{cldf_obj}{a CLDF object}
+}
+\value{
+a dataframe
+}
+\description{
+Returns a table of the foreign keys in a CLDF dataset.
+}
+\examples{
+o <- cldf(system.file("extdata/huon", "cldf-metadata.json", package = "rcldf"))
+get_foreign_keys(o)
+}
