+++ NAMESPACE
@@ -11,6 +11,7 @@ export(get_cache_dir)
 export(get_details)
 export(get_dir_size)
 export(get_filename)
+export(get_foreign_keys)
 export(get_separators)
 export(get_table_from)
 export(get_tablename)
