+++ paper/paper.bib
@@ -11,7 +11,6 @@
   issn         = {2052-4463},
   url          = {http://www.nature.com/articles/sdata2018205}
 }
-
 @book{Glottolog,
   title        = {Glottolog 5.2},
   author       = {Hammarstr{\"o}m, Harald and Forkel, Robert and Haspelmath, Martin and Bank, Sebastian},
@@ -22,16 +21,15 @@
   url          = {https://glottolog.org/},
   urldate      = {2025-09-15}
 }
-
 @book{Concepticon,
   title        = {CLLD Concepticon 3.4.0},
   year         = 2025,
   publisher    = {Max Planck Institute for Evolutionary Anthropology},
   address      = {Leipzig},
+  doi          = {10.5281/zenodo.14923561},
   url          = {https://concepticon.clld.org/},
   editor       = {List, Johann Mattis and Tjuka, Annika and Blum, Frederic and Kučerová, Alžběta and Ugarte, Carlos Barrientos and Rzymski, Christoph and Greenhill, Simon and Forkel, Robert}
 }
-
 @misc{CLTSDS,
   title        = {CLTS. Cross-Linguistic Transcription Systems},
   author       = {Johann-Mattis List and Cormac Anderson and Tiago Tresoldi and Christoph Rzymski and Robert Forkel},
@@ -42,7 +40,6 @@
   copyright    = {Creative Commons Attribution 4.0 International},
   keywords     = {cldf:Generic, linguistics, reference catalog, transcription system}
 }
-
 @article{Anderson_2018,
   title        = {A cross-linguistic database of phonetic transcription systems},
   author       = {Anderson, Cormac and Tresoldi, Tiago and Chacon, Thiago and Fehn, Anne-Maria and Walworth, Mary and Forkel, Robert and List, Johann-Mattis},
@@ -57,16 +54,15 @@
   issn         = {2449-7525},
   url          = {http://dx.doi.org/10.2478/yplm-2018-0002}
 }
-
 @manual{R,
   title        = {R: A Language and Environment for Statistical Computing},
   author       = {{R Core Team}},
   year         = 2025,
   address      = {Vienna, Austria},
+  doi          = {10.32614/r.manuals},
   url          = {https://www.R-project.org/},
   organization = {R Foundation for Statistical Computing}
 }
-
 @manual{csvwr,
   title        = {csvwr: Read and Write CSV on the Web (CSVW) Tables and Metadata},
   author       = {Gower, Robin},
@@ -75,7 +71,6 @@
   url          = {https://CRAN.R-project.org/package=csvwr},
   note         = {R package version 0.1.7}
 }
-
 @article{Lexibank,
   title        = {Lexibank, a public repository of standardized wordlists with computed phonological and lexical features},
   author       = {List, Johann-Mattis and Forkel, Robert and Greenhill, Simon J. and Rzymski, Christoph and Englisch, Johannes and Gray, Russell D.},
@@ -89,7 +84,6 @@
   doi          = {10.1038/s41597-022-01432-0},
   url          = {https://doi.org/10.1038\%2Fs41597-022-01432-0}
 }
-
 @article{TNG,
   title        = {{TransNewGuinea.org: An online database of New Guinea languages}},
   author       = {Greenhill, Simon J.},
@@ -103,17 +97,16 @@
   url          = {http://dx.doi.org/10.1371/journal.pone.0141563},
   pmid         = 26506615
 }
-
 @book{WALS,
   title        = {{WALS Online}},
   year         = 2013,
   publisher    = {Max Planck Institute for Evolutionary Anthropology},
[…]
