### 2026-05-18 77fec8e
try to fix vignette on windows
