### 2026-05-18 3006d19
use tempdir instead of R_user_dir by default
