### 2026-05-20 aed8317
adding State of Field section
