+++ paper/paper.bib
@@ -362,3 +362,31 @@
   editor       = {Berez-Kroeker, Andrea L. and McDonnell, Bradley  and Koller, Eve and Collister, Lauren B.},
   abstract     = {Computational phylogenetics is a relatively recent branch of historical linguistics that uses quantitative techniques to investigate the history of related languages. As the classical comparative method is less explicit on the techniques for constructing phylogenies of language families (see discussion in Jacques & List 2019), such a new approach can complement traditional techniques for sub-grouping based on shared innovations (Ross & Durie 1996).}
 }
+
+
+@Manual{lingtypology,
+   title = {lingtypology: easy mapping for Linguistic Typology},
+   author = {George Moroz},
+   year = {2017},
+   doi = {10.32614/CRAN.package.lingtypology},
+   url = {https://CRAN.R-project.org/package=lingtypology},
+}
+
+@Article{glottospace,
+    title = {glottospace: R package for the geospatial analysis of linguistic and cultural data},
+    author = {Sietze Norder and Laura Becker and Hedvig Skirgård and Leonardo Arias and Alena Witzlack-Makarevich and Rik {van Gijn}},
+    journal = {Journal of Open Source Software},
+    year = {2022},
+    volume = {7},
+    number = {77},
+    pages = {4303},
+    url = {glottospace: R package for language mapping and geospatial analysis of linguistic and cultural data},
+}
+
+@Manual{glottoTrees,
+    title = {{glottoTrees}: Phylogenetic trees in Linguistics.},
+    author = {Erich R. Round},
+    year = {2026},
+    note = {R package version 0.1.13},
+    url = {https://github.com/erichround/glottoTrees},
+  }
