+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Package: rcldf
 Type: Package
 Title: Read Linguistic Data in the Cross Linguistic Data Format (CLDF)
-Version: 1.6.1
+Version: 1.6.2
 Authors@R: 
     person(given = c("Simon", "J."), family = "Greenhill", email = "simon@simon.net.nz", role = c("aut", "cre"))
 Maintainer: Simon J. Greenhill <simon@simon.net.nz>
