### 2026-06-30 acdd17e
rebuild Rd files
