+++ man/get_cldf_colname.Rd
@@ -1,0 +1,30 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/get_cldf_colname.R
+\name{get_cldf_colname}
+\alias{get_cldf_colname}
+\title{Get the Column Name for a CLDF Property}
+\usage{
+get_cldf_colname(cldf_obj, table_conformsto, property)
+}
+\arguments{
+\item{cldf_obj}{A cldf object.}
+
+\item{table_conformsto}{Character string. The short CLDF table type to search,
+e.g. \code{"LanguageTable"} or \code{"FormTable"}.}
+
+\item{property}{Character string. The short CLDF property name to look up,
+e.g. \code{"latitude"}, \code{"longitude"}, or \code{"name"}.}
+}
+\value{
+A character string with the column name, or \code{NULL} if not found.
+}
+\description{
+Looks up the actual column name used in a CLDF table for a given CLDF property,
+by matching against the \code{propertyUrl} field in the dataset metadata.
+This allows code to work correctly regardless of what column names a dataset
+uses for standard CLDF properties (e.g. \code{latitude} vs \code{Latitude}).
+}
+\examples{
+cldfobj <- cldf(system.file("extdata/huon", "cldf-metadata.json", package = "rcldf"))
+get_cldf_colname(cldfobj, "LanguageTable", "latitude")
+}
