+++ man/default_dialect.Rd
@@ -1,12 +1,8 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/csvwr_overrides.R
-\docType{data}
 \name{default_dialect}
 \alias{default_dialect}
 \title{CSVW default dialect}
-\format{
-An object of class \code{list} of length 13.
-}
 \usage{
 default_dialect
 }
@@ -17,4 +13,3 @@ a list specifying a default csv dialect
 The \href{https://w3c.github.io/csvw/metadata/#dialect-descriptions}{CSVW Default Dialect specification}
 described in \href{https://specs.frictionlessdata.io/csv-dialect/}{CSV Dialect Description Format}.
 }
-\keyword{datasets}
