+++ NAMESPACE
@@ -8,6 +8,7 @@ export(cldf)
 export(datasets)
 export(datatype_to_type)
 export(get_cache_dir)
+export(get_cldf_colname)
 export(get_details)
 export(get_dir_size)
 export(get_filename)
