### 2026-06-30 b3ee73f
use a better default for plot_languages
