+++ R/get_foreign_keys.R
@@ -1,5 +1,10 @@
 #' Returns a table of the foreign keys in a CLDF dataset.
 #'
+#' Returns both explicit CSVW \code{foreignKeys} and implicit foreign keys
+#' derived from CLDF reference properties (e.g. \code{languageReference}).
+#' Per the CLDF specification, a column whose \code{propertyUrl} is a CLDF
+#' reference property is equivalent to an explicit foreign key constraint.
+#'
 #' @param cldf_obj a CLDF object
 #' @return a dataframe
 #' @export
@@ -7,14 +12,17 @@
 #' o <- cldf(system.file("extdata/huon", "cldf-metadata.json", package = "rcldf"))
 #' get_foreign_keys(o)
 get_foreign_keys <- function(cldf_obj) {
+    CLDF_BASE <- "http://cldf.clld.org/v1.0/terms.rdf#"
+
     empty <- data.frame(
         SourceTable=character(), SourceColumn=character(),
         DestinationURL=character(), DestinationTable=character(),
         DestinationColumn=character(), stringsAsFactors=FALSE
     )
-    results <- lapply(cldf_obj$metadata$tables, function(table) {
+
+    # --- Explicit CSVW foreignKeys ---
+    explicit_results <- lapply(cldf_obj$metadata$tables, function(table) {
         fks <- table$tableSchema$foreignKeys
-        # Map over each foreign key within that table
         lapply(fks, function(fk) {
             col_ref <- if (length(fk$columnReference) > 0) fk$columnReference[[1]] else NULL
             ref_col <- if (length(fk$reference$columnReference) > 0) fk$reference$columnReference[[1]] else NULL
@@ -33,6 +41,70 @@ get_foreign_keys <- function(cldf_obj) {
             )
         })
     })
-    result <- do.call(rbind, unlist(results, recursive=FALSE))
+    explicit <- do.call(rbind, unlist(explicit_results, recursive=FALSE))
+    if (is.null(explicit)) explicit <- empty
+
+    # --- Implicit foreign keys from CLDF reference properties ---
+    # Per CLDF spec, a propertyUrl of e.g. languageReference is equivalent to
+    # an explicit FK to the LanguageTable's ID column.
+    REF_TO_TABLE <- c(
+        languageReference        = "LanguageTable",
+        parameterReference       = "ParameterTable",
+        codeReference            = "CodeTable",
+        formReference            = "FormTable",
+        cognatesetReference      = "CognatesetTable",
+        contributionReference    = "ContributionTable",
+        entryReference           = "EntryTable",
+        senseReference           = "SenseTable",
+        mediaReference           = "MediaTable",
+        exampleReference         = "ExampleTable",
+        borrowingReference       = "BorrowingTable",
+        partialCognateReference  = "PartialCognateTable"
+    )
+
+    # Build lookup: short table type -> list(url, id_col)
+    table_info <- list()
+    for (t in cldf_obj$metadata$tables) {
+        conformsto <- t[["dc:conformsTo"]]
+        if (is.null(conformsto) || !startsWith(conformsto, CLDF_BASE)) next
+        ttype <- sub(CLDF_BASE, "", conformsto, fixed=TRUE)
+        cols <- t$tableSchema$columns
+        id_col <- "ID"
+        if (!is.null(cols) && !is.null(cols$propertyUrl)) {
+            id_match <- cols$name[!is.na(cols$propertyUrl) & cols$propertyUrl == paste0(CLDF_BASE, "id")]
+            if (length(id_match) > 0) id_col <- id_match[[1]]
+        }
+        table_info[[ttype]] <- list(url=t$url, id_col=id_col)
+    }
+
+    implicit_results <- lapply(cldf_obj$metadata$tables, function(table) {
[…]
