### 2026-06-30 5aebd96
fix property extraction in #53
