### 2026-07-02 9df002a
fix issues noticed in review
