+++ paper/paper.md
@@ -34,14 +34,14 @@ Cross-Linguistic Data Formats [CLDF, @Forkel2018] is a standardized data
 format designed to handle cross-linguistic and cross-cultural datasets. 
 There are currently almost [300 CLDF datasets
 available](https://zenodo.org/search?q=CLDF&f=resource_type%3Adataset&l=list&p=1&s=10&sort=bestmatch),
-representing many of the major publically available linguistic and cultural datasets. 
+representing many of the major publicly available linguistic and cultural datasets. 
 These datasets contain a wide variety of different types of data from the
 world's languages and cultures including catalogues of linguistic metadata, 
 to lexical word lists, grammatical features, phonetic information, 
 geographical locations and areas, as well as religious and cultural traits (Table 1).
 
 CLDF provides a consistent specification and package format ([https://cldf.clld.org/](https://cldf.clld.org/)). This format is a lightweight data-package
-containing one or more data tables containing tabular data in 'CSV on the Web' format (CSVW)
+comprised of one or more data tables containing tabular data in 'CSV on the Web' format (CSVW)
 following the World Wide Web Consortium (W3C) recommendations for [Tabular Data and Metadata](https://www.w3.org/TR/tabular-data-model/). These tables are described and connected by a metadata file in [Javascript Object Notation](http://json.org/) (JSON) format that details the datatypes
 in the CSVW files and how they are related to each other, and to global 'reference catalogs' of 
 linguistic and cultural information [@Forkel2018]. 
@@ -90,7 +90,7 @@ Table 1: Examples of CLDF Datasets showing the dataset, the type of data it cont
 
 Recently a number of _R_ packages have emerged to support analysis of cross-linguistic and cross-cultural data indicating a growing need for R infrastructure for comparative analysis.
 For example, `lingtypology` [@lingtypology] provides an interface to Glottolog and a handful of other CLDF datasets, 
-`glottospace` [@glottospace] provides geographical mapping tools, and `glottoTrees` [@glottoTrees] provides the taxonomic trees frol [@Glottolog].
+`glottospace` [@glottospace] provides geographical mapping tools, and `glottoTrees` [@glottoTrees] provides the taxonomic trees from [@Glottolog].
 
 However, these packages have two important limitations. First, they typically bundle a fixed set of pre-processed datasets, with ad-hoc parsing code tailored to each release. This limitation means they do not generalise to the broader ecosystem of CLDF datasets. More critically, they require ongoing manual maintenance every time an upstream dataset is updated or restructured or else risk distributing stale out-of-date data.
 
@@ -101,21 +101,21 @@ The rcldf package provides the missing link that solves these issues: A general-
 
 # Software Design:
 
-The decision to build a bespoke package for CLDF datasets rather than reuse existing packages was driven by a number of key friction points in the existing functionality in R [@R]. `rcldf` builds on existing R packages [e.g. @csvwr, @jsonlite, @readr, @vroom, @bib2df] and extends them to reduce the risk of analysis bugs and enhance the findability, re-use, and aggregation of CLDF datasets.
+The decision to build a bespoke package for CLDF datasets rather than reuse existing packages was driven by a number of key friction points in the existing functionality in R [@R]. `rcldf` builds on existing R packages [e.g. @csvwr; @jsonlite; @readr; @vroom; @bib2df] and extends them to reduce the risk of analysis bugs and enhance the findability, re-use, and aggregation of CLDF datasets.
 
 **1. Reducing the risk of potential bugs.**
[…]
