### 2026-07-02 4fb7f07
adding CONTRIBUTING.md
