### 2026-07-03 91383c2
make zenodo fetch safer
