### 2026-07-03 8765c89
fix valuetable/cognatetable
