### 2026-07-03 1a09aa3
add details on get_table_from
