### 2026-07-21 5b6cd6e
Update README.md to use pak instead of deprecated devtools
