### 2026-07-25 fcd4deb
linting pass
