+++ tests/testthat/test_as.cldf.wide.R
@@ -15,8 +15,8 @@
 #
 # Should Become:
 #
-#     ID    Parameter_ID Name             Description                Name.ParameterTable      Description.ParameterTable  URL    Area
-# 1 1A-1  1A           Small            A small thing              Consonant Inventories    PD                          http.. Phon..
+#     ID    Parameter_ID Name             Description                Name.ParameterTable      ...
+# 1 1A-1  1A           Small            A small thing              Consonant Inventories    PD
 # 2 1A-2  1A           Moderately small a moderately small thing   Consonant Inventories    PD
 # 3 1A-3  1A           Average          an average thing           Consonant Inventories    PD
 # 4 1A-4  1A           Moderately large a moderately large thing   Consonant Inventories    PD
@@ -25,87 +25,87 @@
 
 
 test_that("test as.cldf.wide", {
-    expect_error(as.cldf.wide('x', ''), "'object' must inherit from class cldf")
+    expect_error(as.cldf.wide("x", ""), "'object' must inherit from class cldf")
 
     df <- cldf(
         system.file("extdata/examples/wals_1A_cldf/StructureDataset-metadata.json", package = "rcldf")
     )
     expect_error(as.cldf.wide(df), "argument \"table\" is missing, with no default")
-    expect_error(as.cldf.wide(df, 'foo'), "Invalid table foo")
+    expect_error(as.cldf.wide(df, "foo"), "Invalid table foo")
     expect_error(as.cldf.wide(df, NA), "Need a table to expand")
 
     # languages has nothing to join, so should just return the same
-    expect_equal(as.cldf.wide(df, 'LanguageTable'), df$table$LanguageTable)
+    expect_equal(as.cldf.wide(df, "LanguageTable"), df$table$LanguageTable)
 
     # now test CodeTable joining ParameterTable:
-    df$tables$ParameterTable$Description <- 'PD'  # set this so it's not NA
+    df$tables$ParameterTable$Description <- "PD"  # set this so it's not NA
 
     # codes joins parameters
-    df.codes <- as.cldf.wide(df, 'CodeTable')
+    df_codes <- as.cldf.wide(df, "CodeTable")
     # should have all codes rows
-    expect_equal(nrow(df.codes), 5)
+    expect_equal(nrow(df_codes), 5)
     cols <- c(
         # from CodeTable:
-        'ID', 'Parameter_ID', 'Name', 'Description',
+        "ID", "Parameter_ID", "Name", "Description",
         # from ParameterTable:
-        'Authors', 'Url', 'Area',
-        'Description.ParameterTable', 'Name.ParameterTable'
+        "Authors", "Url", "Area",
+        "Description.ParameterTable", "Name.ParameterTable"
     )
     for (col in cols) {
-        expect_equal(length(df.codes[[col]]), 5)
+        expect_equal(length(df_codes[[col]]), 5)
     }
 
-    expect_true(all(df.codes$Url == 'http://wals.info/feature/1A'))
-    expect_true(all(df.codes$Authors == 'Ian Maddieson'))
+    expect_true(all(df_codes$Url == "http://wals.info/feature/1A"))
+    expect_true(all(df_codes$Authors == "Ian Maddieson"))
 
-    expect_true(all(df.codes$Description.ParameterTable == df$tables$ParameterTable$Description))
-    expect_true(all(df.codes$Description == df$tables$CodeTable$Description))
+    expect_true(all(df_codes$Description.ParameterTable == df$tables$ParameterTable$Description))
+    expect_true(all(df_codes$Description == df$tables$CodeTable$Description))
 
     # values has everything
[…]
