+++ R/relabel.R
@@ -3,4 +3,4 @@
 #' @param column the tablename.
 #' @param table the tablename.
 #' @return A string of "column.table"
-relabel <- function(column, table) { paste0(column, '.', table) }
+relabel <- function(column, table) paste0(column, ".", table)

[…]
