+++ paper/paper.md
@@ -126,7 +126,7 @@ Third, CLDF datasets are normally stored in a 'long' format where multiple varia
 
 # Research Impact Statement: 
 
-Recent years have seen a major quantitative turn towards large-scale quantitative analysis in linguistics [@Kortmann2021] and humanities [@McGillivray2022]. Cross-Linguistic Data Formats [CLDF, @Forkel2018] has rapidly become a very common standard for producing, releasing, and distributing a rich array of cultural data. CLDF has been suggested as a best-practice approach for distributing data alongside research publications [e.g. @Tresoldi2022]. Publicly available datasets in this format have rapidly increased with currently around 300 datasets online. However, the archival database nature of CLDF is often at conflict with how users want to interact with and analyse data, and the flexibility of CSV formats means that there are many potential bugs waiting to happen with naïve parsing and loading of data. In addition, CLDF datasets are scattered across various websites, dataset repositories, and in journal supplementary materials, requiring users to know how to find, download, unarchive, and load these datasets.
+Recent years have seen a major quantitative turn towards large-scale quantitative analysis in linguistics [@Kortmann2021] and humanities [@McGillivray2022]. Cross-Linguistic Data Formats [CLDF, @Forkel2018] has rapidly become a very common standard for producing, releasing, and distributing a rich array of cultural data. CLDF has been suggested as a best-practice approach for distributing data alongside research publications [e.g. @Tresoldi2022]. Publicly available datasets in this format have rapidly increased with more than 200 datasets online. However, the archival database nature of CLDF is often at conflict with how users want to interact with and analyse data, and the flexibility of CSV formats means that there are many potential bugs waiting to happen with naïve parsing and loading of data. In addition, CLDF datasets are scattered across various websites, dataset repositories, and in journal supplementary materials, requiring users to know how to find, download, unarchive, and load these datasets.
 
 The `rcldf` package resolves these issues by, first, providing metadata-aware loading capabilities that easily load datasets from wherever they may be found, and providing a catalogue of common datasets to make this even easier. This functionality will enable users to readily access and re-use any of these datasets for their own work within the `R` language. Second, the package provides automatic functionality to identify column types, standardize missing data sentinels, and resolve compressed archives like NEXUS and BibTeX. These capabilities will ensure reproducible workflows free from parsing bugs common in ad-hoc CSV imports. Finally, `rcldf` provides a number of tools to explore datasets, from describing the metadata schema, to plotting of data onto maps, to allow users to rapidly understand the dataset they are exploring.
 
