### 2026-07-25 cc323f6
revise tidy/wide/long issue
