+++ paper/paper.md
@@ -121,7 +121,7 @@ First, `rcldf` supports loading CLDF files from not just local sources but websi
 
 Second, `rcldf` contains  functions for automatically loading the CLDF reference catalogs that describe the languages [Glottolog @Glottolog], lexical concepts [Concepticon @Concepticon] and phonetic transcriptions [Cross Linguistic Transcription Systems, @CLTSDS, @Anderson_2018]. These reference catalogs are the key mechanisms which enable datasets to be connected to each other by, for example, aggregating the datasets that contain the Glottolog 'Glottocode' language identifier for a particular language e.g. New Zealand Māori [maor1246](https://glottolog.org/resource/languoid/id/maor1246), or finding datasets with a particular lexical concept of interest.
 
-Third, CLDF datasets are normally stored in a 'long' format where multiple variables are included in the same column matching common database normalisation practices [@Forkel2018]. However, often in statistical analysis users prefer a 'wide' or 'tidy' format [@TidyData] where each variable is contained in its own column, and each observation is a row. To facilitate this, `rcldf` contains tools to convert the 'long' CLDF tables into 'wide' formats while resolving the foreign keys into expanded columns into one data frame.
+Third, CLDF datasets are normally stored in a 'long' format where multiple variables are included in the same column matching common database normalisation practices [@Forkel2018]. However, often in statistical analysis users prefer a 'wide' format [@TidyData] where each variable is contained in its own column, and each observation is a row. To facilitate this, `rcldf` contains tools to converts the 'long' or 'tidy' CLDF tables into 'wide' formats while resolving the foreign keys into expanded columns into one data frame.
 
 
 # Research Impact Statement: 
