+++ paper/paper.md
@@ -20,7 +20,7 @@ bibliography: paper.bib
 
 # Summary
 
-Cross-Linguistic Data Formats (CLDF) is a standardized data format becoming increasingly common for storing
+Cross-Linguistic Data Formats (CLDF) is an increasingly common standardized data format for storing
 and distributing a wide range of comparative linguistic, cultural, ethnographic, geographic, and religious
 data. The `rcldf` package provides a lightweight _R_ toolkit for loading and reading CLDF files from both
 local and remote sources. The package facilitates analysis with _R_ by providing a number of convenience
@@ -37,8 +37,8 @@ available](https://zenodo.org/search?q=CLDF&f=resource_type%3Adataset&l=list&p=1
 representing many of the major publicly available linguistic and cultural datasets. 
 These datasets contain a wide variety of different types of data from the
 world's languages and cultures including catalogues of linguistic metadata, 
-to lexical word lists, grammatical features, phonetic information, 
-geographical locations and areas, as well as religious and cultural traits (Table 1).
+lexical word lists, grammatical features, phonetic information, 
+and geographical locations and areas, as well as religious and cultural traits (Table 1).
 
 CLDF provides a consistent specification and package format ([https://cldf.clld.org/](https://cldf.clld.org/)). This format is a lightweight data-package
 comprised of one or more data tables containing tabular data in 'CSV on the Web' format (CSVW)
@@ -47,7 +47,7 @@ in the CSVW files and how they are related to each other, and to global 'referen
 linguistic and cultural information [@Forkel2018]. 
 
 CLDF is rapidly becoming a standard framework for storing cultural data as it provides a simple, 
-reliable data format that facilitate the storage, sharing, and re-use for these data. The format
+reliable data format that facilitates the storage, sharing, and re-use for these data. The format
 has been used both as the native format for newly released datasets [e.g. @Lexibank; @Lexibank2; @Grambank], as well as to republish legacy datasets after 'retrostandardizing' them for modern
 use [@Forkel2024].
 
@@ -88,7 +88,7 @@ Table 1: Examples of CLDF Datasets showing the dataset, the type of data it cont
 
 ## State of the field
 
-Recently a number of _R_ packages have emerged to support analysis of cross-linguistic and cross-cultural data indicating a growing need for R infrastructure for comparative analysis.
+Recently a number of _R_ packages have emerged to support analysis of cross-linguistic and cross-cultural data, indicating a growing need for R infrastructure for comparative analysis.
 For example, `lingtypology` [@lingtypology] provides an interface to Glottolog and a handful of other CLDF datasets, 
 `glottospace` [@glottospace] provides geographical mapping tools, and `glottoTrees` [@glottoTrees] provides the taxonomic trees from [@Glottolog].
 
@@ -99,7 +99,7 @@ Second, because these packages largely bypass the CSVW/JSON metadata in a CLDF p
 The rcldf package provides the missing link that solves these issues: A general-purpose package for loading and parsing any CLDF dataset. Dataset-specific metadata is incorporated and handled correctly, and the dataset is exposed in a form ready for easy analysis in R. This package should help shift the maintenance burden from per-dataset wrappers to a single shared layer, and open the full CLDF ecosystem to _R_ users.
 
 
-# Software Design:
+# Software design
 
 The decision to build a bespoke package for CLDF datasets rather than reuse existing packages was driven by a number of key friction points in the existing functionality in R [@R]. `rcldf` builds on existing R packages [e.g. @csvwr; @jsonlite; @readr; @vroom; @bib2df] and extends them to reduce the risk of analysis bugs and enhance the findability, re-use, and aggregation of CLDF datasets.
 
@@ -110,31 +110,31 @@ First, `rcldf` is metadata aware, and uses the metadata JSON file that is part o
 
[…]
