### 2026-08-06 5bd4461
Add date to DESCRIPTION file
Added a date entry to the DESCRIPTION file so that `citation()` gives a complete bibTeX entry
