+++ paper/paper.bib
@@ -389,4 +389,12 @@
     year = {2026},
     note = {R package version 0.1.13},
     url = {https://github.com/erichround/glottoTrees},
-  }
+}
+  
+@misc{ELCat,
+  author = {Endangered Languages, Catalogue},
+  date = {2025},
+  title = {University of Hawaiʻi at Mānoa},
+  url = {http://www.endangeredlanguages.com},
+  language = {mg}
+}
