### 2026-08-11 1ad3219
fix doi warnings x 2
