### 2026-08-11 fcd1abf
try to make table better
