### 2026-08-11 152ad2d
Revert "fix doi warnings x 2"
This reverts commit 1ad32193605114c95e3efa315e77379d08e253eb.
