+++ paper/paper.bib
@@ -208,7 +208,7 @@
   year         = 2020,
   publisher    = {},
   address      = {},
-  doi          = {10.5281/zenodo.17433568},
+  doi          = {https://doi.org/10.5281/zenodo.17433568},
   url          = {https://ewave-atlas.org/},
   editor       = {Bernd Kortmann and Kerstin Lunkenheimer and Katharina Ehret}
 }
@@ -380,7 +380,7 @@
     volume = {7},
     number = {77},
     pages = {4303},
-    doi = {10.21105/joss.04303}
+    url = {glottospace: R package for language mapping and geospatial analysis of linguistic and cultural data},
 }
 
 @Manual{glottoTrees,
