### 2026-08-18 #10811 @RichardLitt
@SimonGreenhill Talked to Anna. Let's go with this:

Gary Holton (ed.). Catalogue of Endangered Languages. 2026. University of Hawaiʻi at Mānoa. http://www.endangeredlanguages.com/

Can you make the adjustments?
