+++ paper/paper.md
@@ -62,26 +62,26 @@ A full `vignette` is provided with the R package showing an example analysis and
 | Dataset                                                                                  | CLDF |
 |:-----------------------------------------------------------------------------------------|:-----|
 | **Metadata**                                                                             |      |
-| [Glottolog](https://glottolog.org) [@Glottolog]                                          |  [10.5281/zenodo.18840967](https://doi.org/10.5281/zenodo.18840967)         |
-| [EndangeredLanguages.com](https://www.endangeredlanguages.com) [@ELCat]                  |  [10.5281/zenodo.13946786](https://zenodo.org/records/13946786)         |
+| [Glottolog](https://glottolog.org) [@Glottolog]                                          |  [1🔗](https://doi.org/10.5281/zenodo.18840967)         |
+| [EndangeredLanguages.com](https://www.endangeredlanguages.com) [@ELCat]                  |  [2🔗](https://zenodo.org/records/13946786)         |
 | **Lexicon**                                                                              ||
-| [Lexibank](https://lexibank.clld.org/) [@Lexibank]                                       |  [10.5281/zenodo.5227817](https://doi.org/10.5281/zenodo.5227817)      |
-| [TransNewGuinea.org](https://transnewguinea.org) [@TNG]                                  |  [10.5281/zenodo.14162587](https://zenodo.org/records/14162587)         |
-| [Indo-European Cognate Relationships](https://iecor.clld.org/) [@Anderson_2025]          |  [10.5281/zenodo.8089434](https://doi.org/10.5281/zenodo.8089434)      |
+| [Lexibank](https://lexibank.clld.org/) [@Lexibank]                                       |  [3🔗](https://doi.org/10.5281/zenodo.5227817)      |
+| [TransNewGuinea.org](https://transnewguinea.org) [@TNG]                                  |  [4🔗](https://zenodo.org/records/14162587)         |
+| [Indo-European Cognate Relationships](https://iecor.clld.org/) [@Anderson_2025]          |  [5🔗](https://doi.org/10.5281/zenodo.8089434)      |
 | **Grammatical**                                                                          ||
-| [Grambank](https://grambank.clld.org) [@Grambank]                                        |   [10.5281/zenodo.7844558](https://zenodo.org/records/7844558)          |
-| [AUTOTYP](https://www.isle.uzh.ch/en/DLL/Databases-and-Methods/AUTOTYP.html) [@AUTOTYP]  |   [10.5281/zenodo.7976754](https://zenodo.org/records/7976754)          |
-| [The World Atlas of Language Structures](https://wals.info) [@WALS]                      |   [10.5281/zenodo.13950591](https://zenodo.org/records/13950591)         |
-| [Electronic World Atlas of Varieties of English](https://ewave-atlas.org) [@EWAVE]       |   [10.5281/zenodo.3712132](https://zenodo.org/records/3712132)          |
+| [Grambank](https://grambank.clld.org) [@Grambank]                                        |   [6🔗](https://zenodo.org/records/7844558)          |
+| [AUTOTYP](https://www.isle.uzh.ch/en/DLL/Databases-and-Methods/AUTOTYP.html) [@AUTOTYP]  |   [7🔗](https://zenodo.org/records/7976754)          |
+| [The World Atlas of Language Structures](https://wals.info) [@WALS]                      |   [8🔗](https://zenodo.org/records/13950591)         |
+| [Electronic World Atlas of Varieties of English](https://ewave-atlas.org) [@EWAVE]       |   [9🔗](https://zenodo.org/records/3712132)          |
 | **Phonetic**                                                                             ||
[…]
