+++ paper/paper.bib
@@ -391,10 +391,15 @@
     url = {https://github.com/erichround/glottoTrees},
 }
   
+% Gary Holton (ed.). Catalogue of Endangered Languages. 2026. University of Hawaiʻi at Mānoa. http://www.endangeredlanguages.com/
 @misc{ELCat,
-  author = {Endangered Languages, Catalogue},
-  date = {2025},
-  title = {University of Hawaiʻi at Mānoa},
-  url = {http://www.endangeredlanguages.com},
-  language = {mg}
+  editor    = {Holton, Gary},
+  title     = {Catalogue of Endangered Languages},
+  year      = {2026},
+  publisher = {University of Hawai{\textquoteleft}i at M{\={a}}noa},
+  address   = {Honolulu},
+  url       = {http://www.endangeredlanguages.com/},
+  urldate   = {2026-08-19}
+
 }
+
