### 2026-08-18 e6216ec
attempt at adding a emoji for link out did not work, revert
