+++ paper/paper.bib
@@ -394,7 +394,7 @@
 % Gary Holton (ed.). Catalogue of Endangered Languages. 2026. University of Hawaiʻi at Mānoa. http://www.endangeredlanguages.com/
 @misc{ELCat,
   editor    = {Holton, Gary},
-  title     = {Catalogue of Endangered Languages},
+  title     = {Catalogue of {E}ndangered {L}anguages},
   year      = {2026},
   publisher = {University of Hawai{\textquoteleft}i at M{\={a}}noa},
   address   = {Honolulu},
