### 2026-08-19 #10811 @RichardLitt
@SimonGreenhill Thanks. Can you make a new release and deposit it on Zenodo and share the DOI and version here? I'm following the editorial guide here (https://joss.readthedocs.io/en/latest/editing.html#post-review). @openjournals/sbcs-eics Do we have a document that explains this process to authors?
