### 2026-08-19 #10811 @RichardLitt
@SimonGreenhill Did you fix `https://doi.org/10.5281/zenodo.17433568 is INVALID because of 'https://doi.org/' prefix`?
