+++ paper/paper.bib
@@ -208,7 +208,7 @@
   year         = 2020,
   publisher    = {},
   address      = {},
-  doi          = {https://doi.org/10.5281/zenodo.17433568},
+  doi          = {10.5281/zenodo.17433568},
   url          = {https://ewave-atlas.org/},
   editor       = {Bernd Kortmann and Kerstin Lunkenheimer and Katharina Ehret}
 }
