### 2026-08-19 #10811 @editorialbot
:wave: @openjournals/sbcs-eics, this paper is ready to be accepted and published.
