### 2026-08-21 #10811 @RichardLitt
@samhforbes The other authors only had one contribution each to the main repo, or were editors/reviewers on this paper. I think solo author would be fine here.
